reallocarray: error on 0 sizes
OpenBSD reallocarray(3) will return a special pointer that will cause a
segfault when used if size or nmemb is 0.

Since the compat reallocarray(3) is a wrapper around realloc(3), the
behaviour is a little more complicated on other OS'es. On Linux:

    ptr: NULL, size: 0 -> malloc(NULL, 0) : either ENOMEM or pointer
    ptr: NULL, size: size -> malloc(NULL, size)

    ptr: non-NULL, size: 0 -> free(ptr)
    ptr: non-NULL, size: size -> realloc(ptr, size)

The OpenBSD and FreeBSD man pages for realloc(3) do not specify the
behaviour for realloc(ptr, 0) when ptr is non-NULL.